+++ crates/just-lsp-wasm/Cargo.toml
@@ -15,4 +15,5 @@ just-lsp = { version = "0.4.2", path = "../.." }
 lsp-types = "0.94.1"
 serde = { version = "1.0.228", features = ["derive"] }
 serde-wasm-bindgen = "0.6.5"
+typeshare = "1.0.5"
 wasm-bindgen = "0.2.118"
