+++ www/src/lib/types.ts
@@ -1,9 +1,14 @@
-export interface SyntaxNode {
-  id: number;
-  type: string;
-  text: string;
-  startPosition: { row: number; column: number };
-  endPosition: { row: number; column: number };
-  childCount: number;
-  children: SyntaxNode[];
+/*
+ Generated by typeshare 1.13.3
+*/
+
+export interface Diagnostic {
+	end_character: number;
+	end_line: number;
+	id: string;
+	message: string;
+	severity: string;
+	start_character: number;
+	start_line: number;
 }
+
